requirements.txt
asyncio
Jinja2
requests
numpy
Pillow
fastapi
pydantic
transformers[torch]
uvicorn[standard]
gunicorn
python-multipart
scikit-learn==1.0.2
tensorflow==2.10.0